+++ R/GetWriteR.R
@@ -1,6 +1,36 @@
 # look out for two versions of GetWriteR() second one is while the exe file is broken.
 # look out for two versions of GetWxPython27(), the second is experimental and uses a Python shell command command to pull wxPython
 
+
+TestPython = function(){
+if(nchar(Sys.which("python"))>0){
+VersionString = system2("python", "--version", stdout=TRUE, stderr=TRUE)
+message("Your system is using ", VersionString)
+return(TRUE)
+}
+else{
+message("Python cannot be seen on your system.\nIf it is installed, then you may need to ensure your system settings are correct.\n")
+return(FALSE)
+}
+}
+
+TestWX = function(){
+if(TestPython()){
+      if (interactive()) {
+        if (.Platform$OS.type == "windows") {
+shell(paste("python", system.file("Python/TestWX.py", package="BrailleR")))
+message("One of two things just happened; either\n1. a window popped up using the necessary toolkit, which you should have closed, and you're ready to go with WriteR\nor,\n2. some error messages were printed as R output, in which case the WXPython toolkit needs to be installed.\n")
+        } else {
+          warning(
+              "This function is for users running R under the Windows operating system.\n")
+        }
+      } else {
+        warning("This function is meant for use in interactive mode only.\n")
+      }
+}
+}
+
+
 GetWriteR =
     function(UseGitHub = TRUE) {
       if (interactive()) {
@@ -66,61 +96,6 @@ GetPandoc =
     }
 
 
-
-GetRStudio =
-    function() {
-      if (interactive()) {
-        if (.Platform$OS.type == "windows") {
-          if (requireNamespace("BrailleR")) {
-            if (requireNamespace("installr")) {
-              message(
-                "This command will download a file and save it to your hard drive.\n")
-              installr::install.RStudio(
-                download_dir = getOption("BrailleR.Folder"),
-                keep_install_file = TRUE)
-              message(
-                "The installer file has been added to your MyBrailleR folder.")
-              message(
-                "You can delete it at any time, but that will not uninstall the application.")
-            }
-          }
-        } else {
-          warning(
-              "This function is for users running R under the Windows operating system.\n")
-        }
-      } else {
-        warning("This function is meant for use in interactive mode only.\n")
-      }
-      return(invisible(NULL))
-    }
-
-
-
-Get7zip =
-    function() {
-      if (interactive()) {
-        if (.Platform$OS.type == "windows") {
-          if (requireNamespace("BrailleR")) {
-            if (requireNamespace("installr")) {
-              message(
-                "This command will download a file and save it to your hard drive.")
-              installr::install.7zip(download_dir = getOption("BrailleR.Folder"),
-                                   keep_install_file = TRUE)
-              message(
-                "The installer file has been added to your MyBrailleR folder.")
-              message(
-                "You can delete it at any time, but that will not uninstall the application.")
-            }
-          }
-        } else {
-          warning(
-              "This function is for users running R under the Windows operating system.\n")
-        }
-      } else {
-        warning("This function is meant for use in interactive mode only.\n")
-      }
-      return(invisible(NULL))
-    }
 
 GetPython27 =
     function() {
@@ -253,37 +228,6 @@ GetWxPython27 =
       return(invisible(Success))
     }
 
-
-GetCygwin =
-    function(x64=TRUE) {
-      if (interactive()) {
-        if (.Platform$OS.type == "windows") {
-          if (requireNamespace("BrailleR")) {
-            if (requireNamespace("installr")) {
-              message(
-                "This command will download a file and save it to your hard drive.\n")
-bit=32
-if(x64){
-              if(installr::is.x64()) bit=64
-              }
-      installr::install.cygwin(bit=bit,
-                download_dir = getOption("BrailleR.Folder"),
-                keep_install_file = TRUE)
-              message(
-                "The installer file has been added to your MyBrailleR folder.")
-              message(
-                "You can delete it at any time, but that will not uninstall the application.")
-            }
-          }
-        } else {
-          warning(
-              "This function is for users running R under the Windows operating system.\n")
-        }
-      } else {
-        warning("This function is meant for use in interactive mode only.\n")
-      }
-      return(invisible(NULL))
-    }
 
 GetWxPython3 =
     function() {