+++ NEWS.md
@@ -1,5 +1,8 @@
 # BrailleR 0.30.0
 - making ready for upload to CRAN
+- attempting to remove all references to the reticulate package
+- added TestPython() and TestWX() functions which test system setup.
+
 
 # BrailleR 0.29.5
 - work to remove Python 2.7 from package files